creat-getRanking and approveRanking and publishRanking

+++ SWP_Hackathon/src/main/java/com/hackathon/dto/ranking/RankingResponseDTO.java
@@ -4,6 +4,7 @@ import com.hackathon.entity.enums.ParticipantStatus;
 import com.hackathon.entity.enums.RoundStatus;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,7 +13,7 @@ import java.util.List;
 @Builder
 public class RankingResponseDTO {
     private Integer participantId;
-    private double totalScore;
+    private BigDecimal totalScore;
     private Integer rank;
     private String teamName;
     private ParticipantStatus status;
