+++ SWP_Hackathon/src/main/java/com/hackathon/entity/TeamParticipant.java
@@ -5,8 +5,10 @@ import com.hackathon.entity.enums.ParticipantStatus;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Setter
@@ -25,8 +27,8 @@ public class TeamParticipant {
     private ParticipantStatus status;
 
     @Column(name = "Total_Score")
-    private double totalScore;
-
+//    private double totalScore;
+    private BigDecimal totalScore;
     @Column(name = "Rank")
     private Integer rank;
